Parse output of /proc/cmdline as part of OATT check and catch device access exception when drivers aren't sourced
Summary: Host side reduction unit test made more robust to running on machines that don't have drivers sourced or lacking OATT support by checking the value of `/proc/cmdline`.

Test Plan:
Unit test should run on all setups, all of the tyan development machines are setup with OATT so I haven't tried this on a machine without OATT.

If it's still flaky then the last resort is to create a tiny poplar program in the unit test that uses remote buffers and check whether or not it runs.

Reviewers: jamesn, #popart, george

Reviewed By: jamesn, #popart

Subscribers: george

Maniphest Tasks: T15990

Differential Revision: https://phabricator.sourcevertex.net/D21051