Add type hints, docstrings and tidy front end.
Summary:
A bunch of tidying documentation and type hints for the python front end. No functionality changes.

This makes it easier for the user to see expected types and info about the expected values for the different inputs, especially on a good IDE.

It was Friday and I was on a lot of painkillers, so there may be a few spelling errors.

Test Plan: All tests pass.

Reviewers: #popart, jamesn

Reviewed By: #popart, jamesn

Subscribers: jamesn

Maniphest Tasks: T16118

Differential Revision: https://phabricator.sourcevertex.net/D21301