Cpp Hardware tests functionality
Summary:
This is 1/2 for the task T15569 done just for cpp files. 95% of the changes are just swapping out the `createCpuModelDevice` for the `createTestDevice(TEST_TARGET)` function in testdevice.hpp`. There are a couple of tests that swap out devices within the test which I will need to adapt for this.

This mirrors the functionality in poplar from the `TestDevice.hpp`. Hopefully this should mean we can run only hardware tests / all tests on hardware /ony  CPU tests etc.

The python side should be the same, I will probably add to the test util file a function to get the device based on the cmake arguments.

Test Plan: All pass.

Reviewers: #popart, jamesn

Reviewed By: #popart, jamesn

Subscribers: jamesn

Maniphest Tasks: T15569

Differential Revision: https://phabricator.sourcevertex.net/D21439