Add support for writing to and reading from external TensorProto data for initializers
Summary:
For very large onnx models it is not desirable to store the data for initializer TensorData inside the onnx model. This is because it impacts performance of serialising the model, and operating on the serialised model. Protobuf sets a hard 2GB limit on message size.

To allow for this, TensorProto has an `external_data` field that allows you to specify that it is stored externally, with a number of options stored as a map.

To support this, I have made changes in 3 main places:
- Methods to the Builder and GraphTransformer classes, `saveInitializersExternally`: Can now convert a model a lot of initializer data to a smaller one with externally stored tensors
- `onnxutil::getConstData` changed to support TensorProtos with external data : Can now allows create an IR from an onnx model with external data.
- `Session::modelToHost` changed to now overwrite the binary external data files when writing out the rest of the onnx model.

Test Plan:
Tests added to:
Python unit tests added to a new file: tests/popart/external_tensorproto_data_test.py

All tests pass

Reviewers: #popart, jamesn

Reviewed By: #popart, jamesn

Subscribers: jamesn

Maniphest Tasks: T13994

Differential Revision: https://phabricator.sourcevertex.net/D21725