Python test functionality for hardware only tests
Summary:
Ready to review!

Sort of regret taking this on...

This allows functionality for the python tests to be run on different devices, as well as running hardware only tests, similar as for the cpp tests.

Unfortunately this was a lot more complex than the cpp tests, as I had to account for ctest running and pytest. For ctest a test is made for each variant / device and the test run with the environment variable `TEST_TARGET`. The user will need to specify these in CMakeLists.txt with a `VARIANT`. Python tests now have to be added explicitly per test, ideally creating separate tests for Hw and non-Hw tests for the same functionality.

For running pytest separately, you no longer have the CMakeLists variant to go off, so the user adds the `@tu.requires_ipu` decorator. This adds the TEST_TARGET env variable if not already overridden by ctest.  This is a bit awkward, not sure if there is a better way of doing this.

I will create a follow up task to tidy up some of the test functionality, perhaps merging `test_util` `op_tester` and `test_session`. Turning the create_test_device into a pytest fixture would be good as well.

Test Plan: All tests pass with .test.sh and pytest.

Reviewers: #popart, jamesn

Reviewed By: #popart, jamesn

Subscribers: evan, jamesn

Maniphest Tasks: T15569

Differential Revision: https://phabricator.sourcevertex.net/D21466