Add session option, and optimization pass at the end of Ir::createBackwards() to decompose gradient summation into a series of adds in the optimal order
Summary:
Fixes T15398

See diagrams in the code for a visual explanation of the transform applied by the new session option `decomposeGradSum`

Note: This diff depends on the new scheduler (D19826) to ensure that:
- the tree of new `Add` ops is constructed optimally
- the `Add` ops are scheduled as early as possible (if that is the best thing to do for sum liveness).

With D19826, it is asserted that, in typical cases when this new session option is expected to be used (e.g. see cpp unit test) that:
- the order of adds is for gradient partials closest to the loss to those furthest away from the loss.
- the adding of grad partials is scheduled as early as possible once the grad partials become live. Note: this is not a hard constraint. It is left up to the scheduler, but is verified in the new cpp unit test.

Test Plan:
All tests pass.

Added a python test to show numerical equivalence with the new session option turned on.
Added cpp test to check scheduling and order of addition tree

Reviewers: #popart, fabiant, jamesn

Reviewed By: #popart, fabiant, jamesn

Subscribers: fabiant, jamesn, matejk

Maniphest Tasks: T15398

Differential Revision: https://phabricator.sourcevertex.net/D20523