Parse env variables to option flag passed to GCL allReduce
Summary: Ref T15704.

Test Plan: I have not managed to run resnet with the diff as it seems that resnet does not work with the popart dev branch. I have done is to run an earlier version on master. It is not exactly the same code though.

Reviewers: #popart, evan, fabiant

Reviewed By: #popart, fabiant

Maniphest Tasks: T15704

Differential Revision: https://phabricator.sourcevertex.net/D25484