Order ops in builder.h.gen.
Summary:
Previously, the order of the contents of generated files was non-deterministic.
This caused large diffs when running `gen_operators.py` and it was difficult to see the actual changes.

In this diff, I manually sorted the contents of the generated files. They are functionally unchanged in this diff.
I made the changes to `gen_operators.py` after manually sorting the generated files, and checked that running `gen_operators.py` produced no changes.

Now running `gen_operators.py` should only change the generated files if there is new information to add to them.

Reviewers: matthewha, evan, #popart

Reviewed By: evan, #popart

Maniphest Tasks: T17985

Differential Revision: https://phabricator.sourcevertex.net/D23403