Add ONNX_NAMESPACE alias for onnx throughout.
Summary:
I think this is what the ticket asked for...?

As per the task, MSFT want to be able to change the onnx namespace to something different to enable building against the onnx runtime. This was what I could see was the simplest solution, MSFT cange change the cmake variable, wheras for everyone else, the api remains the same.

Happy to try alternatives if anyone has an opinion.

Reviewers: #popart, davidn, simonl, jamesn

Reviewed By: #popart, jamesn

Maniphest Tasks: T16085

Differential Revision: https://phabricator.sourcevertex.net/D23156