Use pipeline stage information to determine required pipeline depth
Summary:
The check in pipeline.cpp that batchesPerStep or accumulationFactor is large enough for the pipeline depth used numIPUs. This is inaccurate, since it is possible to create a pipelined model with > 1 pipeline stage per IPU. Changed the check to use numPipelineStages.

The other changes are non-functional : tidied some confusing code in the PipelineInfo class.

Fixes T18151

Reviewers: #popart, jamesn

Reviewed By: #popart, jamesn

Subscribers: jamesn

Maniphest Tasks: T18151

Differential Revision: https://phabricator.sourcevertex.net/D23463