Standalone diff of explicit recomputation taken from T9357
Summary:
Explicit recomputation is a technique that trades computation for memory, by performing on-demand explicit recomputation during the backwards step.
It can be enabled through the session options.

Test Plan:
Two tests added: `explicit_recomputation_test.cpp` and `explicitrecomputation.py`.
The former checks if the **explicitRecomputation** settings is correctly set (updated) and exists in the IR.
The latter is a more end-to-end test that compares the result with and without **explicitRecomputation** turned on/off.

Reviewers: evan, matthewha, george, #popart, jamesn

Reviewed By: evan, #popart, jamesn

Subscribers: jamesn, fabiant

Tags: #popart

Maniphest Tasks: T17050

Differential Revision: https://phabricator.sourcevertex.net/D22164