Remove test that relies on startTileMultiplier
Summary:
Ref T5208. This option has been removed.

The fact that this option had error handling in Poplar was being used to implictly test that the conv options were being set by popART. Because there isn't an option to replace this with that has the same characteristic I have removed the test.

As an aside I think this is an example of [[ https://www.hyrumslaw.com | Hyrum's Law ]] and that mocking would be better solution for this sort of test.

Reviewers: george, jamesn, #popart

Reviewed By: jamesn, #popart

Maniphest Tasks: T5208

Differential Revision: https://phabricator.sourcevertex.net/D24454