Change enable.sh.in to use CMAKE_INSTALL_LIBDIR and LD_PATH_VARNAME
Summary:
These variables are provided by cmake's GNUInstallDirs package and
are already used by poplar. The purpose is to allow the enable.sh
to be generated correctly for Centos. CMAKE_INSTALL_LIBDIR is
already used in willow so the build libraries and script will
match up.

Test Plan: Run manually and check if it works on Centos, Ubuntu and Mac OX X

Reviewers: #popart, evan, jamesn, fabiant, matthewha, george, davidn

Reviewed By: #popart, davidn

Subscribers: davidn

Maniphest Tasks: T19061

Differential Revision: https://phabricator.sourcevertex.net/D25036