Split the download_dataset "test" into many "tests"
Summary:
By having multiple "tests", if the download step fails without
verbose output (such as buildbots), it is easier to identify
which steps failed.

"tests" which did not need to run, e.g. if the dataset is already
downloaded are marked as skipped to distinguish between a completion
of a task.

Reviewers: #popart, jamesn

Reviewed By: #popart, jamesn

Maniphest Tasks: T19953

Differential Revision: https://phabricator.sourcevertex.net/D25107