Extend PopART GraphTransformer to better handle unsupported ONNX Tensor types
Summary:
Adds some methods in the graphTransformer to convert unsupported ONNX tensor types to some that are actually supported by PopART:
* int8, int16, uint8, uint16, int64  ------> int32
* double  ------> float32
* double  ------> half
* bfloat16  ------> float32

solves T16679

// code duplication removal task, see T20159

Test Plan:
Adds a unit test that creates an onnx graph with input and output of the type to convert and a single constant node in it which is initialized to some know values. It then runs the proto through the graphtransformer method that is appropriate and checks that the data is preserved, that it is in the right data field and that tensor shape is preserved.

The script is called `type_casting_test.py` and has been added to the `CMakeList.txt`.

After building, to run:

```
POPART_LOG_LEVEL=INFO ./test.sh popart -R Cpu_default_type_casting_test -VV
```

Reviewers: jamesn, evan, #popart

Reviewed By: jamesn, #popart

Tags: #popart

Maniphest Tasks: T16679

Differential Revision: https://phabricator.sourcevertex.net/D24055