Fix ELU and SELU implementations
Summary:
The bug was as simple as a pe::_1 vs pe::_2 and did not get picked
up by the test as the sign coming in was correct.

The tests are also varied to use a smaller range since otherwise
the test fails when using an identity loss and sum, simply because
the numbers are very small leading to numerical error.

Reviewers: #popart, fabiant

Reviewed By: #popart, fabiant

Maniphest Tasks: T20230

Differential Revision: https://phabricator.sourcevertex.net/D25415