Make Optimiser tensors FLOAT
Summary:
It can't be this simple .. ?

If the LR etc scalars are constant, then they are stored internally as C++ floats, then fed into the scaledAdd / subtract .

If they are non-constant then tensors are created for each of them, in this case making them float, rather than the data type of the weights. When it comes to do the scaled add `auto swd1Val = vu_op.initSwd1.val();` will be a float (as it's just taking the info from the float tensors created earlier.

Momentum is more awkward, upcasting the toUpdate tensor, multiplying, then downcasting. I'm sure there is a better way to do this.

Then if the `swd1Val` cannot be cast as FLOAT16 to match the other 2:
```
 *      Where the tensors \p A, \p B are of type half and a \p scaleB
 *      is provided as a float or a tensor of type float, it is possible to
 *      to implement the scaledAddTo in half precision if \p scaleB can be cast
 *      to half precision with acceptable accuracy.  Otherwise full precision
 *      arithmetic can be used internally, but at the cost of speed.
 *      Floating point arithmetic will be selected if the relative error in
 *      casting is greater than the relative tolerance. \n
 *      Only applies to ScaledAddTo with scaleB.
```

So if swd1Val is a 'proper' float, the calculation will be done in FLOAT.

Reviewers: davidn, oskarb, #popart, jamesn

Reviewed By: oskarb, #popart, jamesn

Tags: #popart

Maniphest Tasks: T13366

Differential Revision: https://phabricator.sourcevertex.net/D25992