Change flatten->reshape in mergevarupdates
Summary:
This replaces the flatten in place ops with reshape in place ops in the mergevarupdates transform. This avoids conflicts later on when flatten in place should have been replaced by reshape in place, but the transform is run too late.

I changed most of the naming refs to reshape, which wasn't necessary and might be confusing, so I have tried to explain this in the code.

Test Plan: Test pass, I'll let the buildbot run as I'm unsure what else this could touch.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #popart, jamesn, markk, evan

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #popart, evan

Maniphest Tasks: T22105

Differential Revision: https://phabricator.sourcevertex.net/D27598