Register Constant op and similar with the OpManager
Summary: Rather than create dummy ops, I have just created dummy entires in the OpManager. The factory functions will throw an error if called. This is ok, as the factory functions should not be called as long as the ops are handled by `OnnxConstExprUtil`.

Test Plan: Added a test to check that the ops are present in `getSupportedOps`.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Subscribers: evan, simonl

Maniphest Tasks: T22341

Differential Revision: https://phabricator.sourcevertex.net/D28158