Change the generated code for the opset11 version of the constant op.
Summary:
The builder api for the constant op no longer worked in opset 11.
As of opset11, the constant op has 2 attributes, but only 1 should be set.
The builder api required both to be set and when a builder generated model using this op was run through the onnx model checker, it would throw an error.

Test Plan: I've added a test to make sure that models generated with the constant op, pass the onnx model checker.

Reviewers: #popart, jamesn, evan, matthewha, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: #popart, jamesn, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T17985

Differential Revision: https://phabricator.sourcevertex.net/D23456