Changed espilon in nll loss (rebased)
Summary:
Changes the epsilon value based on the probabilities type to avoid underflow/overflow when running in fp16, for which the previous eps was out of range.
Also changed the addition of the eps in the forward pass to a max to avoid introducing bias.

Solves T22699

Test Plan:
All tests pass.

Attached script also no longer has NaNs
{F985307}

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Subscribers: evan

Maniphest Tasks: T22699

Differential Revision: https://phabricator.sourcevertex.net/D28755