Fix int64 torch models
Summary:
Issue with these was the commented `convertINT64ToINT32` code in `torchWriter.py`, however using this did present another problem that need to be handled.

When converting all int64 tensors to int32, the shape input to the reshapes was converted too. However in the onnx spec this is invalid as the shape tensor must be int64. I could have selectively converted the onnx model to avoid this, however I felt that would have been too much work. Instead, it only converts to int32 where we have added int64 tensors as part of nllloss labels etc. This will mean that a combination of nllloss and reshape will case this script to fail.

This isn't an issue with popART, just that blanket converting to int32 in general is not a good idea.

Test Plan: Will let the buildbots have a go.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Maniphest Tasks: T20968

Differential Revision: https://phabricator.sourcevertex.net/D28999