Re-implement SliceGrad, making re-use of inplace Pad by modified class inheritance
Summary:
1
BEFORE: the inheritance structure for Ops
PadInplace :  BasePadOp
Pad :         BasePadOp
SliceGradOp : Op

AFTER:
PadInplace :  BasePadOp
Pad :         BasePadOutplaceOp
SliceGradOp : BasePadOutplaceOp

2
BEFORE: PadSum pattern (to convert Sum of Pads to concats where
appropriate) is  only applied to the Pad class.

AFTER: PadSum pattern (to convert Sum of Pads to concats where appropriate) is
applied to the BasePadOutplaceOp class, so SliceGradOp can benefit too.

3
BEFORE: the inheritance structure for Opxs
PadOpx :        Opx
PadInplaceOpx : Opx
SliceGradOpx :  Opx

AFTER:
PadOpx :        BasePadOpx
PadInplaceOpx : BasePadOpx
SliceGradOpx is removed, SliceGradOp creates a PadOpx directly.

4
BEFORE:
Search for propitious padding layout is applied only to SliceGradOpx.
If it is found, the propitious layout is cloned, and then at every iteration of
the program, the clone is reset with a Copy program to 0.

AFTER:
Search for propitious layout is applied to BasePadOpx, so any PadOp can benefit.
There is a task to generalize this (T22334).
The cloning has changed slightly too - instead of cloning all of the propitious
Tensor, only the required padding part has it's tile mapping copied, and it is
copied by a constant Tensor which means we do not need the Copy program. There
is task to expose this mapping approach to poplibs (T22336)

5
BEFORE: PadInplace::modifies does not return the empty Region, this bug is fixed.

Test Plan:
4 tests added.
- confirm Pad numerically correct with default layout
- confirm Pad numerically correct with cloned layout
- confirm SliceGrad becomes PadInplace
- confirm PadSum applied appropriately to SliceGrad
- confirm Tensor Mapping is identical in Slice's in and SliceGrad's out when propitious layout found

Reviewers: #popart, evan, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, george, fabiant

Reviewed By: #popart, evan, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T22007

Differential Revision: https://phabricator.sourcevertex.net/D27716