Support constant padding zero-dim tensors (and thus taking zero-dim slices, as their gradients are pads)
Summary:
Handle zero-dim constant mode tensors in PadX (which is the grad OpX of a slice).

In zero-dim case, pad with `::NONE` mapping method, then manually map with a `popart::LinearMapper`.

For edge or reflect mode paddings, it does not make sense to pad a zero-dim constant, so we let it fail in poplibs.

Test Plan:
Test out-of-bounds slicing in `operator_tests.py`, running the `op_tester` in training mode. Tests follow what onnx's own tests assert, as well as checking the gradient tensor.

https://github.com/onnx/onnx/blob/master/onnx/backend/test/case/node/slice.py

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, evan, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, matthewha

Maniphest Tasks: T21626

Differential Revision: https://phabricator.sourcevertex.net/D29201