Add a new factory function which allows more control over op creation.
Summary:
The current factory function returns a `std::unique_ptr<OP>` to the created op. The graph then moves ownership of the op pointer to itself, connects the ops inputs and outputs, and calls `Op::setup` on the new op.

The new factory function returns an `Op*` to the created op. The factory function is responsible for moving the ownership of the op pointer to the graph, and connecting the ops inputs and outputs. The graph is still responsible for calling `Op::setup` on the new op.

Note:

This diff changes the api of OpManager::registerOp.

Old API:
```
  static void registerOp(const OperatorIdentifier &opid,
                         const OpDefinition &details,
                         bool isPublic,
                         OpFactoryFunc func);
```
New API:
```
  static void registerOp(const OpInfo &opInfo);
```

This should not be a problem, as custom ops in user code should be using the `OpCreator` interface.

Test Plan: The `Pad11` factory function has been converted to the new op, and the pad tests in `operators_tests.py` are still passing. These tests definitely check the differences between `Pad11` and previous versions of `Pad`.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #popart, jamesn, evan, matthewha

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #popart, matthewha

Subscribers: matthewha

Maniphest Tasks: T21732

Differential Revision: https://phabricator.sourcevertex.net/D29282