Throw error if stepio tensor not in inputs
Summary:
Not sure of the flow of this one. But it just throws an error if a stepio tensorId is in the maingraph tensors , but not in the onnx inputs.

Should I move this to a different area ? Or refactor the stepio_size_assertion as it's doing a lot now?

Test Plan: Added a test for the error

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, evan

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Maniphest Tasks: T23122

Differential Revision: https://phabricator.sourcevertex.net/D29553