Ensure attributes set for primitive typed defaults
Summary:
Without this commit, the builder does not set attributes when the
value is equal to the default. This causes ONNX type and shape
inference to fail in some circumstances, e.g. for RandomUniform
This commit allows attributes to be set to default values in the
builder if the type is an int or float.

This commit also fixes the Poptorch issue: T25476.

Test Plan: Popart tests and reproduced in T25476

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan, jamesn, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan

Subscribers: simonl

Maniphest Tasks: T25891

Differential Revision: https://phabricator.sourcevertex.net/D31096