Fix issue in the op to identity pattern.
Summary:
The `OpToIdentity` pattern could replace the `GatherOp` with an `IdentityOp`, which could result in the error:
```
default fwdRegMap not valid : should be specialised for 102 (ai.graphcore.IdentityInplace:1)
```

This was due to the replacement `IdentityOp` having multiple inputs.

The fix this diff makes is to disconnect all but the first input of the op that is about to be replaced by an identity op.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Maniphest Tasks: T23410

Differential Revision: https://phabricator.sourcevertex.net/D31072