Don't divide by replicationFactor in the loss output when using Mean reduction
Summary:
[API change]
Stop the output loss value when using replication + ReductionType.Mean from being divided by the replicationFactor. This is unintuitive as when returning the loss tensor, the anchor includes a replicationFactor dim. So the reduction hasn't happened over that dimension.

Note: this is not to be confused with the gradients of the loss which will still be divided by the replicationFactor.

Improve consistency when using ReductionType.Mean w.r.t the global batch size (1/3)

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Subscribers: jamesn, fabiant

Maniphest Tasks: T25727

Differential Revision: https://phabricator.sourcevertex.net/D31887