Don't assert on invalid memory proportion use
Summary: Right now if the user adds an available memory attribute on any op that isn't a conv or matmul PopART throws an exception. This is needless because it is an implementation detail so it can be safely ignored silently. Having this also limits how users can interact with the API

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Maniphest Tasks: T26394

Differential Revision: https://phabricator.sourcevertex.net/D31901