Support odd scale factor.
Summary: The scale factor tensor was used to calculate the output shape, but then the `output_dimension / input_dimension` was the scale factor actually used in the resize operation. This matched the behaviour in pytorch 1.5.1. Pytorch 1.6.0 changed this behaviour.  The onnx specification doesn't mention recomputing the scale factor. This diff makes resize match the behaviour of pytorch 1.6.0 and is likely the intended onnx spec behaviour.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, anthonyb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Maniphest Tasks: T26403

Differential Revision: https://phabricator.sourcevertex.net/D32479