Made it so that exceptions in callbacks are passed through poplar and caught by popart's pybind11.
Summary:
The problem was different than initially anticipated. I thought the problem was to do with typeids of exception
classes not matching when in the callstack we have popart -> poplar -> popart via callback but this is not the case.
The actual problem was that some callback functions in popart were declared as noexcept whilst, under some conditions,
throwing an exception. This is not allowed in C++ and, when you do throw an exception in a noexcept method, the runtime
aborts your progam. This is what was happening and this is why pybind11 didn't catch the exception. I have fixed this
by removing the noexcept qualifiers.

Test Plan:
* Re-enable unit test that test the condition above (np_memory_layout_test.py::test_np_memory_layout_add_input_tensor_pystepiocallback)
* Check no other tests fail.

Reviewers: jamesn, evan, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: jamesn, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T23055

Differential Revision: https://phabricator.sourcevertex.net/D32784