Fixed issue with IStepIO data order between session runs
Summary: The problem was that there is an additional `prefetch` callback from poplar at the end of a session run. Because of how PopART needs to split data from one source into multiple sets of callbacks for poplar, this `prefetch` resulted in a call to the underlying data source `inComplete` method, which advances the data, ready for the next call. However, this next call never happens and the session ends in a state with the underlying IStepIO having advanced more than it should have. To combat this, this diff only calls the downstream `inComplete` if either 1) it really needs to advance because we need the element for the next replica or 2) there is a matching upstream `inComplete` call from poplar.

Test Plan: Unit tests.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, george, jamesn

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Maniphest Tasks: T27637

Differential Revision: https://phabricator.sourcevertex.net/D33239