Fix build failures to do with copying popart/popart/python/popart.
Summary:
The problem was that where previously one dir copy we now do one for each module. In certain timings of parallelization this is causing build problems. I guess we were asking to copy the same directory multiple times at the same time, and it didn't like that. Really, we only need to copy this directory once so I have changed it to copy based on the `popart_core` target, much like how it used to be. Although the buildbots passing will not tell us if we fixed it it may tell us if we broke it in some other way.

NOTE: This is probably not the right fix, see D33416

Test Plan:
* Buildbots
* @bjornf to confirm it works.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, bjornf

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, bjornf

Subscribers: bjornf, jorgec

Maniphest Tasks: T27824, T27362

Differential Revision: https://phabricator.sourcevertex.net/D33412