Add support for shaped dropout to popart
Summary: This diff defines a custom `ShapedDropout` operator that targets the `poprand` library function of the same name.  This operator requires a `shape` parameter that is used to define the shape of the dropout mask so that strongly correlated features in the input data set can be preserved.  This is supported in pytorch using `torch.nn.Dropout2d` and `torch.nn.Dropout3d` as well as in tensorflow using the `noise_shape` parameter (T18496)

Test Plan: Added various tests that check the behaviour of the RNG sampling in different scenarios to confirm the expected independence / repeatability of the computation.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T24836

Differential Revision: https://phabricator.sourcevertex.net/D33064