Remove references to complete poplar types from DeviceManager header
Summary:
We remove references to complete types from poplar. This hides the dependency: allows the user to include this file and compile without adding poplar include dirs.

Note, however, due to the fundamental design of this class, the user will still need to create poplar types themselves to interact with it. So will probably be including and linking poplar anyway. But this is a first step towards having a public-private separation.

---

 - Include `<map>` in device manager (as we refer to it, was compiling by luck before).
 - Remove `#include <poplar/DeviceManager>` and forward declare poplar types used.
 - Change field `DeviceInfo::flags` from `const poplar::OptionFlags` to a `const unique_ptr<const poplar::OptionFlags>`. This retains the const semantics.
 - Change the getter and constructor accordingly, moving their definitions into the cpp.

I'm only changing a private field. The API of the public constructor and getter are unchanged.

Test Plan: hope the buildbots like it

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, anthonyb, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha, simonl, hatemh

Maniphest Tasks: T27680

Differential Revision: https://phabricator.sourcevertex.net/D33273