Change test to check for just one occurrence of expected available memory proportion
Summary:
I did a poplibs change to cache the plans generated from the FWD pass. This causes the BWD and WU passes to not log the specific line expected by the test.
However the line does still get logged for the FWD pass. This change is to check for just one occurrence of the line in the poplibs log.

Ref T23238

Reviewers: oskarb, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: oskarb, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T23238

Differential Revision: https://phabricator.sourcevertex.net/D33837