Avoid lowering a groupnorm op to poplar when its data input has 0 elements
Summary:
A fix for a bit of a contrived example. If an input to a groupnorm has 0 elements (i.e. a dimension of size 0), a segfault is thrown inside `popnn::gn::groupNormStatistics`. Avoid lowering a 0-element-input groupnorm op from popart to poplar by converting it to an identity.

Should this really be fixed in poplar? Yes, probably. Here's the ticket for them: T28462. But this is still a valid fix in the mean time.

Test Plan: python poptorch/tests/torch_nn_test.py test_nn_LayerNorm_1d_empty_elementwise_affine no longer fails with groupnorm segfault

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb, george, jamesn

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Subscribers: jamesn

Maniphest Tasks: T26648

Differential Revision: https://phabricator.sourcevertex.net/D33941